Move SocialMedia dto to contact.py
