+++ backend/dto/source.py
@@ -1,6 +1,7 @@
 from pydantic import Field
 from typing import Optional, List
 from backend.dto.common import PaginatedRequest, RequestDTO
+from backend.dto.contact import UpdateSocialMedia
 
 
 class CreatePartner(RequestDTO):
@@ -11,21 +12,6 @@ class CreatePartner(RequestDTO):
         None, description="Website URL of the partner.")
     slug: Optional[str] = Field(
         None, description="Slug for the partner organization.")
-
-
-class UpdateSocialMedia(RequestDTO):
-    twitter_url: Optional[str] = Field(
-        None, description="Twitter URL of the source.")
-    linkedin_url: Optional[str] = Field(
-        None, description="LinkedIn URL of the source.")
-    facebook_url: Optional[str] = Field(
-        None, description="Facebook URL of the source.")
-    instagram_url: Optional[str] = Field(
-        None, description="Instagram URL of the source.")
-    youtube_url: Optional[str] = Field(
-        None, description="YouTube URL of the source.")
-    tiktok_url: Optional[str] = Field(
-        None, description="TikTok URL of the source.")
 
 
 class UpdatePartner(RequestDTO):
